+++ common/src/main/java/com/selfdot/cobblemonmegas/common/command/permissions/Permission.java
@@ -1,0 +1,11 @@
+package com.selfdot.cobblemonmegas.common.command.permissions;
+
+import net.minecraft.util.Identifier;
+
+public record Permission(String literal, PermissionLevel level) {
+
+    public Identifier getIdentifier() {
+        return new Identifier("selfdot", literal);
+    }
+
+}