+++ common/src/main/java/com/selfdot/cobblemonmegas/common/command/permissions/PermissionLevel.java
@@ -1,0 +1,11 @@
+package com.selfdot.cobblemonmegas.common.command.permissions;
+
+public enum PermissionLevel {
+
+    NONE,
+    SPAWN_PROTECTION_BYPASS,
+    CHEAT_COMMANDS_AND_COMMAND_BLOCKS,
+    MULTIPLAYER_MANAGEMENT,
+    ALL_COMMANDS
+
+}