+++ common/src/main/java/com/selfdot/cobblemonmegas/common/command/permissions/PermissionValidator.java
@@ -1,0 +1,9 @@
+package com.selfdot.cobblemonmegas.common.command.permissions;
+
+import net.minecraft.command.CommandSource;
+
+public interface PermissionValidator {
+
+    boolean hasPermission(CommandSource source, Permission permission);
+
+}