+++ common/src/main/java/com/selfdot/cobblemonmegas/common/command/permissions/VanillaPermissionValidator.java
@@ -1,0 +1,12 @@
+package com.selfdot.cobblemonmegas.common.command.permissions;
+
+import net.minecraft.command.CommandSource;
+
+public class VanillaPermissionValidator implements PermissionValidator {
+
+    @Override
+    public boolean hasPermission(CommandSource source, Permission permission) {
+        return source.hasPermissionLevel(permission.level().ordinal());
+    }
+
+}