+++ fabric/src/main/java/com/selfdot/cobblemonmegas/fabric/CobblemonMegasFabric.java
@@ -3,7 +3,8 @@ package com.selfdot.cobblemonmegas.fabric;
 import com.selfdot.cobblemonmegas.common.CobblemonMegas;
 import net.fabricmc.api.ModInitializer;
 
-public class CobblemonMegasFabric implements ModInitializer {
+public class
+CobblemonMegasFabric implements ModInitializer {
 
     @Override
     public void onInitialize() {