+++ forge/src/main/kotlin/com/selfdot/cobblemonmegas/forge/CobblemonMegasForge.kt
@@ -2,7 +2,10 @@ package com.selfdot.cobblemonmegas.forge
 
 import com.selfdot.cobblemonmegas.common.CobblemonMegas
 import com.selfdot.cobblemonmegas.common.DataKeys
+import com.sun.org.apache.xalan.internal.xsltc.compiler.util.Type.Void
+import net.minecraftforge.api.distmarker.Dist
 import net.minecraftforge.common.MinecraftForge
+import net.minecraftforge.fml.DistExecutor
 import net.minecraftforge.fml.common.Mod
 import net.minecraftforge.fml.event.lifecycle.FMLCommonSetupEvent
 import thedarkcolour.kotlinforforge.KotlinModLoadingContext
@@ -18,6 +21,7 @@ class CobblemonMegasForge {
 
     private fun commonSetup(event: FMLCommonSetupEvent) {
         CobblemonMegas.getInstance().onInitialize()
+        DistExecutor.safeRunWhenOn(Dist.DEDICATED_SERVER) { SetPermissionValidatorRunnable() }
     }
 
 }