+++ forge/src/main/kotlin/com/selfdot/cobblemonmegas/forge/SetPermissionValidatorRunnable.kt
@@ -1,0 +1,11 @@
+package com.selfdot.cobblemonmegas.forge
+
+import com.selfdot.cobblemonmegas.common.CobblemonMegas
+import net.minecraftforge.fml.DistExecutor
+
+class SetPermissionValidatorRunnable : DistExecutor.SafeRunnable {
+    override fun run() {
+        CobblemonMegas.getInstance().setPermissionValidator(ForgePermissionValidator())
+    }
+
+}