+++ CHANGELOG.md
@@ -1,5 +1,6 @@
 # 1.1.3
 - Fixed Mega raid bosses in CobblemonRaids being changed to normal form on first battle.
+- Fixed permissions not working on Arclight.
 
 # 1.1.2
 - Fixed crash on start when using Arclight.